fixed error in the Issued event
