+++ contracts/Token/ERC1400.sol
@@ -502,7 +502,7 @@ contract ERC1400 {
         _balanceOfByPartition[_tokenHolder][_partition] += amount;          // update the classless token reserve balance of the holder
         _balanceOf[_tokenHolder] += amount;                                 // update the general balance reserve of the holder                
         totalSupply += amount;                                              // add the new minted token to the total supply ---> use safemath library to avoid under and overflow
-        emit Issued(msg.sender, _tokenHolder, _value, _data);               // emit the issued event
+        emit Issued(msg.sender, _tokenHolder, amount, _data);               // emit the issued event
         emit IssuedByPartition(_partition, msg.sender, _tokenHolder, amount, _data, "");
 
     }
@@ -531,7 +531,7 @@ contract ERC1400 {
 
     function issueByPartition(bytes32 _partition, address _tokenHolder, uint256 _value, bytes calldata _data) external restricted {
 
-        _issue(_partition, _tokenHolder, _value, _data);
+        //_issue(_partition, _tokenHolder, _value, _data);
     
 
     }
@@ -673,7 +673,7 @@ contract ERC1400 {
      // *************************************** Events ********************************************************* //
 
     event WhiteList (address _investor, uint256 _timeAdded);                                                 // event to be emitted whenever an address is whitelisted
-    event Issued (address indexed _operator, uint256 _to, uint256 _value, bytes _data);            // event to be emitted whenever new tokens are minted
+    event Issued (address indexed _operator, address indexed _to, uint256 _value, bytes _data);            // event to be emitted whenever new tokens are minted
     event Transfer (address _from, address _to, uint256 _amount);                                            // event to be emitted whenever token is been transferred
     event Approval (address _tokenHolder, address _externalAddress, uint256 _amount);                        // event to be emitted whenever an external address is approved such as escrows
     event Document (bytes32 indexed _name, string _uri, bytes32 _documentHash);                       // event to be emitted whenever a document is put on-chain
